+++ src/Sulu/Bundle/SnippetBundle/Domain/Event/WebspaceDefaultSnippetModifiedEvent.php
@@ -13,8 +13,9 @@ namespace Sulu\Bundle\SnippetBundle\Domain\Event;
 
 use Sulu\Bundle\EventLogBundle\Domain\Event\DomainEvent;
 use Sulu\Bundle\SnippetBundle\Admin\SnippetAdmin;
+use Sulu\Bundle\SnippetBundle\Document\SnippetDocument;
 
-class DefaultSnippetRemovedEvent extends DomainEvent
+class WebspaceDefaultSnippetModifiedEvent extends DomainEvent
 {
     /**
      * @var string
@@ -26,14 +27,21 @@ class DefaultSnippetRemovedEvent extends DomainEvent
      */
     private $snippetAreaKey;
 
+    /**
+     * @var SnippetDocument
+     */
+    private $snippet;
+
     public function __construct(
         string $webspaceKey,
-        string $snippetAreaKey
+        string $snippetAreaKey,
+        SnippetDocument $snippet
     ) {
         parent::__construct();
 
         $this->webspaceKey = $webspaceKey;
         $this->snippetAreaKey = $snippetAreaKey;
+        $this->snippet = $snippet;
     }
 
     public function getSnippetAreaKey(): string
@@ -41,15 +49,23 @@ class DefaultSnippetRemovedEvent extends DomainEvent
         return $this->snippetAreaKey;
     }
 
+    public function getSnippet(): SnippetDocument
+    {
+        return $this->snippet;
+    }
+
     public function getEventType(): string
     {
-        return 'default_snippet_removed';
+        return 'default_snippet_modified';
     }
 
     public function getEventContext(): array
     {
         return [
             'snippetAreaKey' => $this->snippetAreaKey,
+            'snippetId' => $this->snippet->getUuid(),
+            'snippetTitle' => $this->snippet->getTitle(),
+            'snippetTitleLocale' => $this->snippet->getLocale(),
         ];
     }
